# docs/swagger.yaml
...
openapi: "3.0.0"
info:
  title: Demli API
  version: 1.0.0
  description: >
    Multi-domain consumer API — rent-a-car, travel, hotel, food,
    booking flow (FlowBox), payments, reviews, wishlists.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
...
paths:
...
  /api/rentacar/companies/{id}:
    get:
      tags: [RentACar]
      summary: Get rental company details
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Company details }
  /api/rentacar/cars:
    get:
      tags: [RentACar]
      summary: List cars with filters
      security: []
      parameters:
        - in: query
          name: companyId
          schema: { type: string }
        - in: query
          name: brand
          schema: { type: string }
        - in: query
          name: minPrice
          schema: { type: number }
        - in: query
          name: maxPrice
          schema: { type: number }
      responses:
        "200": { description: Car list }
  /api/rentacar/cars/{id}:
    get:
      tags: [RentACar]
      summary: Get car details
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Car details }

  # ── Travel ──────────────────────────────────────────────────────────────
...
  /api/travel/companies/{id}:
    get:
      tags: [Travel]
      summary: Get travel company details
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Company details }
  /api/travel/tours:
    get:
      tags: [Travel]
      summary: List tours with filters
      security: []
      parameters:
        - in: query
          name: companyId
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: minRating
          schema: { type: number }
        - in: query
          name: startDate
          schema: { type: string, format: date }
        - in: query
          name: endDate
          schema: { type: string, format: date }
        - in: query
          name: name
          schema: { type: string }
      responses:
        "200": { description: Tour list }
  /api/travel/tours/{id}:
    get:
      tags: [Travel]
      summary: Get tour details
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Tour details }
  /api/included-services/{serviceType}:
    get:
      tags: [Travel]
      summary: Get included services by type
      security: []
      parameters:
        - in: path
          name: serviceType
          required: true
          schema: { type: string }
      responses:
        "200": { description: Included services list }

  # ── Hotel ───────────────────────────────────────────────────────────────
  /api/hotel:
    get:
      tags: [Hotel]
      summary: List hotels with filters
      security: []
      parameters:
        - in: query
          name: minPrice
          schema: { type: number }
        - in: query
          name: maxPrice
          schema: { type: number }
        - in: query
          name: starRating
          schema: { type: integer }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: minRating
          schema: { type: number }
        - in: query
          name: name
          schema: { type: string }
      responses:
        "200": { description: Hotel list }
  /api/hotel/{id}:
    get:
      tags: [Hotel]
      summary: Get hotel details
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Hotel details }
  /api/hotel/{id}/rooms:
    get:
      tags: [Hotel]
      summary: List rooms for a hotel
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: roomType
          schema: { type: string }
      responses:
        "200": { description: Room list }

  # ── Food ────────────────────────────────────────────────────────────────
...